refactor: Slider 컴포넌트명과 이름 수정
