+++ src/routes/Home/Slider.tsx
